+++ cmake/CheckCxxCompilerSupport.cmake
@@ -8,11 +8,11 @@
 # provides friendly hints to the user.
 
 if (${CMAKE_CXX_COMPILER_ID} STREQUAL "Clang")
-    if (${CMAKE_CXX_COMPILER_VERSION} VERSION_LESS "3.5.0")
+    if (${CMAKE_CXX_COMPILER_VERSION} VERSION_LESS "3.9.1")
         message(WARNING "
     ### You appear to be using Clang ${CMAKE_CXX_COMPILER_VERSION}, which is known
     ### to be unable to compile Hana. Consider switching to
-    ### Clang >= 3.5.0. If it is already installed on your
+    ### Clang >= 3.9.1. If it is already installed on your
     ### system, you can tell CMake about it with
     ###
     ###     cmake .. -DCMAKE_CXX_COMPILER=/path/to/clang
@@ -43,7 +43,7 @@ elseif (${CMAKE_CXX_COMPILER_ID} STREQUAL "AppleClang")
     ### shipped with Xcode < 6.3. Unfortunately, only Apple's Clang
     ### >= 6.1.0 (shipped with Xcode >= 6.3) can compile Hana.
     ### You should consider updating to Xcode >= 6.3 (requires Yosemite)
-    ### or using a non-Apple Clang >= 3.5.0, which can be installed via
+    ### or using a non-Apple Clang >= 3.9.1, which can be installed via
     ### Homebrew with
     ###
     ###     brew install llvm --with-clang