+++ RELEASE_NOTES.md
@@ -1,4 +1,5 @@
 ## These are the Release notes for the next release of Hana
-- Official support for Xcode 6 and 7 has been dropped. The library should still
-  work with these compilers, however they are not being tested regularly anymore,
-  so they are not officially supported.
+- Official support for Xcode 6, 7 and 8, and LLVM Clang 3.5, 3.6, 3.7, and 3.8
+  has has been dropped. The library should still work with these compilers,
+  however they are not being tested regularly anymore, so they are not
+  officially supported.